local_llm/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
diskcache==5.6.3
distro==1.9.0
filelock==3.16.1
fsspec==2024.9.0
gguf==0.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
Jinja2==3.1.4
jiter==0.6.1
llama_cpp_python==0.3.1
MarkupSafe==3.0.1
mpmath==1.3.0
networkx==3.4.1
numpy==1.26.4
openai==1.51.2
packaging==24.1
protobuf==4.25.5
pydantic==2.9.2
pydantic_core==2.23.4
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
sentencepiece==0.2.0
sniffio==1.3.1
sympy==1.13.3
tokenizers==0.20.1
torch==2.2.2
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
urllib3==2.2.3
